Efficiency
Changed the code so that it stores the info for free games that were pulled in the previous session. Basically if the list of free games pulled includes a game that was pulled before, we get that saved data so we don't have to make API calls again. This greatly improves the speed and efficiency of the program.